fix(test): resolve type conversion and Riverpod 3.x legacy imports in unit tests

+++ lib/features/clipboard_history/presentation/quick_panel_screen.dart
@@ -47,7 +47,8 @@ class _QuickPanelScreenState extends ConsumerState<QuickPanelScreen> {
     final items = ref.read(historyControllerProvider).visibleItems;
     if (items.isEmpty) return;
     setState(() {
-      _selectedIndex = (_selectedIndex + direction).clamp(0, items.length - 1);
+      _selectedIndex =
+          (_selectedIndex + direction).clamp(0, items.length - 1).toInt();
     });
     ref
         .read(historyControllerProvider.notifier)
